weather-api/weather_project/requirements.txt
Django
djangorestframework==3.12.4
requests==2.25.1
django-cors-headers